+++ activerecord/lib/rails/generators/active_record/migration/templates/timestamps.rb
@@ -1,0 +1,5 @@
+class <%= migration_class_name %> < ActiveRecord::Migration
+  def change
+    <%= migration_action %>_timestamps :<%= table_name %>
+  end
+end